use uint32_t to store the # of bytes in a file. Max filesize would be about 4.3gb which is WAY more than enough.

+++ ViciEngine/SimplePacket.hpp
@@ -63,7 +63,7 @@ namespace Networking {
 
 		template <>
 		void add(const std::string& str) {
-			add(str.size());
+			add(static_cast<uint32_t>(str.size()));
 			for (const char& c : str) {
 				_data.push_back(c);
 			}
@@ -111,7 +111,7 @@ namespace Networking {
 
 		template <>
 		inline std::string get() {
-			size_t size = get<size_t>();
+			uint32_t size = get<uint32_t>();
 			std::string str{ _data.data() + _pos, size };
 			_pos += size;
 			return str;
